readme correction
