+++ projects/commons-lib/src/lib/logic/constants/constants.ts
@@ -1,0 +1,1 @@
+export const IMG_BASE = "https://img.freepik.com/vector-premium/no-hay-foto-disponible-icono-vector-simbolo-imagen-predeterminado-imagen-proximamente-sitio-web-o-aplicacion-movil_87543-10615.jpg"
