+++ projects/commons-lib/src/lib/logic/index.ts
@@ -4,3 +4,4 @@ export * from "./state";
 export * from "./interfaces";
 export * from "./interceptors";
 export * from "./guards";
+export * from "./constants";
