+++ projects/home/src/app/containers/payment/payment.component.html
@@ -4,7 +4,7 @@
       <span class="font-semibold text-2xl">Resumen de compra</span>
     </div>
     <div *ngFor="let product of products" class="flex items-center gap-4 border p-4 rounded-lg shadow-sm bg-white">
-      <img [src]="product.thumbnail" alt="{{ product.title }}" class="w-20 h-20 object-cover rounded" />
+      <img [src]="imgBase" alt="{{ product.title }}" class="w-20 h-20 object-cover rounded" />
       <div class="flex flex-col justify-between flex-1">
         <div class="text-lg font-semibold text-gray-800">
           {{ product.title }}
